Move system test to separate file
